+++ gDRplots/NEWS.md
@@ -1,3 +1,7 @@
+## gDRplots 0.0.99 - 2026-01-13
+* improve QC plots for plate
+* add heatmaps for combo metrics data
+
 ## gDRplots 0.0.98 - 2026-01-02
 * update x-axis label for `plot_volcano_assoc_panel`